requirements.txt
streamlit==1.37.0
pandas==2.1.1
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.2
google-generativeai==0.3.2
langchain==0.2.5
langchain-community==0.2.4
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.16.3
fpdf==1.7.2